refactoring artrace -> artraceslits
